requirements.txt

django
django-cors-headers
django-extensions
django-filter
django-storages
djangorestframework
markdown
django-filter
python-dateutil
shortuuid
